Enable strict compiler warnings and treat warnings as errors
Add comprehensive warning flags to build system:
- General: -Wall -Wextra -Wpedantic -Werror -Wshadow -Wcast-align
  -Wunused -Wconversion -Wsign-conversion -Wnull-dereference
  -Wdouble-promotion -Wformat=2 -Wimplicit-fallthrough
- C-specific: -Wstrict-prototypes -Wmissing-prototypes -Wold-style-definition
- C++-specific: -Wnon-virtual-dtor -Woverloaded-virtual

Fix all warnings in source code:
- Cast atomic int32_t values to uint32_t for statistics
- Add explicit casts for signed/unsigned conversions
- Cast bitwise operations to appropriate types
- Mark unused parameters with (void) casts
- Initialize arrays to avoid uninitialized warnings
- Add unsigned literals (U suffix) where needed
- Fix format-nonliteral warnings with pragmas

Build configuration changes:
- Apply warning flags only to project code, not third-party libraries
- Allow -Wno-missing-field-initializers in test code for brevity
- Warnings applied to layered_queue library and all_tests target

Verification:
- Zero warnings, zero errors in build
- All 332 tests pass (100%)
- Strict type safety now enforced throughout codebase

+++ src/drivers/lq_engine.c
@@ -151,7 +151,7 @@ void lq_process_merges(
         int32_t result = -999;  /* Initialize to obviously wrong value for debugging */
         enum lq_status vote_status = LQ_OK;
         
-        int rc = lq_vote(values, valid_count, merge->voting_method, merge->tolerance, &result, &vote_status);
+        int rc = lq_vote(values, valid_count, merge->voting_method, (int32_t)merge->tolerance, &result, &vote_status);
         
         /* Convert lq_status to lq_event_status */
         enum lq_event_status result_status = LQ_EVENT_OK;
@@ -395,7 +395,7 @@ static void *engine_thread_func(void *arg)
             /* Convert sample to event */
             /* In real implementation, route through mid-level drivers */
             events[num_events].source_id = sample.src;
-            events[num_events].value = sample.value;
+            events[num_events].value = (int32_t)sample.value;
             events[num_events].status = LQ_EVENT_OK;
             events[num_events].timestamp = sample.timestamp;
             num_events++;
